ch: create a different file for controller and routes for jobs
